+++ packages/react-native/package.json
@@ -13,6 +13,7 @@
     "@react-navigation/bottom-tabs": "^6.6.0",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/stack": "^6.4.0",
+    "babel-plugin-module-resolver": "^5.0.2",
     "design-system": "workspace:^",
     "nativewind": "^2.0.11",
     "react": "18.2.0",
