+++ packages/react-native/src/pages/Login/Signup.tsx
@@ -1,0 +1,3 @@
+export default function Signup() {
+  return;
+}
